added type 1 and type 2 to pokemonFight calc.c switch case almost complete

+++ Pokemon/Calc.h
@@ -4,4 +4,4 @@
 #include "Util.h"
 #include "Attacks.h"
 
-void mainAttack(PokemonFight *angreifer, PokemonFight *verteidiger,int attackID);
+void mainAttack(PokemonFight *angreifer, PokemonFight *verteidiger,byte attackID);
